PR comments: Superclass
